flv+srt: "smear" captions, that looks more like broadcast 609.
This is a work in progress, but a good point to commit.

This now reads a cue from an SRT file, and instead of adding a bunch of
chained captions to a single frame, it makes a command list and queues
sending the rest of it.

Then, every frame, we send two blocks down the line.

This needs a little extra magic, because timestamps can be out of order in
the FLV.

This fixes some other bugs relating to this program too, and improves
things:

* now only attaches captions to video frames with a later timestamp than the
  previous one.

* cea708_render now adds filler if you don't have a multiple of 5 blocks to
  send.  caption-inspector suggests that this should _actually_ be something
  based on framerate, but this mostly works for now.

* commands_for_frame replaces sei_from_caption_frame; this now acts a bit
  more like broadcast streams.

* don't ENM (clear the non-display memory) when hiding a caption

* remove noisy debug from vtt parse_timestamps

* sei_free has null-check

* sei_dump_messages now dumps h264, instead of h262. probably still wrong